requirements.txt
javalang
requests
python-dotenv
matplotlib
pandas